Fix TIDA comparitor plotting when no reference is available
This is needed to handle the case where no reference histograms
are available


Former-commit-id: 8d0b8d5c38caab58717582b0d3677b560b8963f9

+++ Trigger/TrigAnalysis/TrigInDetAnalysisUser/scripts/TIDArun.sh
@@ -52,52 +52,42 @@ if [ -e TrkNtuple-0000.root ]; then
       REFFILE=$REFDIR/$2
     fi
 
-    if [ -e  $REFFILE ]; then 
+    NOREF=
+
+    if [ ! -e $REFFILE ]; then 
+
+      echo "no reference file $REFFILE - will process without"
+
+      NOREF=--noref
+
+    else
+
       echo "copying $REFFILE" 
 
       LOCALFILE=$(basename $REFFILE)
 
-      # don;t bother to copy it if it already exists in this directory
-      if [ ! -e $LOCALFILE ]; then 
-         cp $REFFILE .
-      fi
+      # don't bother to copy it if it already exists in this directory
+      [ -e $LOCALFILE ] || cp $REFFILE .
+  
+    fi
 
-      ls -l
+    ls -l
 
-      if [ "x$EXPERT" == "x" ]; then 
+    for arg in $*; do 
+       if $(echo "$arg" | grep -q "TIDA.*.dat"); then get_files -data "$arg"; fi 
+       if $(echo "$arg" | grep -q "Test.*.dat"); then get_files -data "$arg"; fi 
+    done
 
-        echo
-        echo "running comparitor " `date`
-        echo
-
-        for arg in $*; do 
-           if $(echo "$arg" | grep -q "TIDA.*.dat"); then get_files -data "$arg"; fi 
-           if $(echo "$arg" | grep -q "Test.*.dat"); then get_files -data "$arg"; fi 
-        done
-
-        TIDAcomparitor.exe $*
-
-      else
-
-        echo
-        echo "running cpucost " `date`
-        echo
-
-        for arg in $*; do 
-           if $(echo "$arg" | grep -q "TIDA.*.dat"); then get_files -data "$arg"; fi 
-           if $(echo "$arg" | grep -q "Test.*.dat"); then get_files -data "$arg"; fi 
-        done
-
-        TIDAcpucost.exe $*
-
-      fi
-
+    if [ "x$EXPERT" == "x" ]; then 
+        echo -e "\nrunning comparitor " $(date) "\n"
+        TIDAcomparitor.exe $* $NOREF
     else
-      echo "no reference file $REFFILE" 
+        echo -e "\nrunning cpucost " $(date) "\n"
+        TIDAcpucost.exe $* $NOREF
     fi
 
     echo "finished postprocessing"
-    ls -l
 
+    ls -l
 
 fi
